+++ psmigrate/migrate.go
@@ -1,10 +1,10 @@
 package psmigrate
 
 import (
-	"github.com/pentops/protostate/internal/pgstore/pgmigrate"
+	"github.com/pentops/protostate/internal/pgstore/psmpg"
 	"github.com/pentops/protostate/psm"
 )
 
 func BuildStateMachineMigrations(specs ...psm.QueryTableSpec) ([]byte, error) {
-	return pgmigrate.BuildStateMachineMigrations(specs...)
+	return psmpg.BuildStateMachineMigrations(specs...)
 }
